Extend demo with an OIDC Identity provider
The demo is extended with a Keycloak instance that can be used as an
identity provider for OpenBao. Furthermore the provisioning of OpenBao
was extended to configure the Keycloak instance as OIDC Identity
Provider.

To test the integration the user `foo` with the password `bar` can be
used in OpenBao.